+++ spec/unit/alf-algebra/support/test_relational.rb
@@ -15,6 +15,7 @@ module Alf
           Matching,
           Minus,
           NotMatching,
+          Page,
           Project,
           Quota,
           Rank,